Only delete `Menu` when empty (#218)
* Only remove Windows start menu directory if empty

* Add install/remove test

* Remove redundant warning for terminal profile

* Only remove Linux directory entry when menu is empty

* Add news item

* Check if menu location exists before iterating

* Update menuinst/platforms/win.py


---------

+++ menuinst/platforms/win.py
@@ -36,8 +36,16 @@ class WindowsMenu(Menu):
         return (self.start_menu_location,)
 
     def remove(self) -> Tuple[os.PathLike]:
-        log.debug("Removing %s", self.start_menu_location)
-        shutil.rmtree(self.start_menu_location, ignore_errors=True)
+        # Only remove if the Start Menu directory is empty in case applications share a folder.
+        menu_location = Path(self.start_menu_location)
+        if menu_location.exists():
+            try:
+                # Check directory contents. If empty, it will raise StopIteration
+                # and only in that case we delete the directory.
+                next(menu_location.iterdir())
+            except StopIteration:
+                log.debug("Removing %s", self.start_menu_location)
+                shutil.rmtree(self.start_menu_location, ignore_errors=True)
         return (self.start_menu_location,)
 
     @property
@@ -344,7 +352,6 @@ class WindowsMenuItem(MenuItem):
 
         if remove:
             if index < 0:
-                log.warning(f"Could not find terminal profile for {name}.")
                 return
             del settings["profiles"]["list"][index]
         else:
